Fix set-face-attribute error
For some reason or other, using font here doesn't work anymore, so we use height
to limit the map-tile-face to less than 16 px. However, the documentation does
say that height controls the attribute in units of 1/10th of a pt, so this may
be a really bad idea that screws things up for you.