Make map window and inventory window if not exist
Since nethack 3.6.6, the nhapi-create-map-window and
nhapi-create-inventory-window functions are called after
nhapi-restore-window-configuration. while it may be possible to write a patch
for this in the startup in the source, writing the function like this is easier.
Now, when nhapi-restore-window-configuration is called, if they do not already
exist, a map buffer and inventory buffer are created using
nhapi-create-map-window and nhapi-create-inventory-window respectively.
nhapi-create-map-window and nhapi-create-inventory-window have been altered so
that if they are called after nhapi-restore-window-configuration, and a map or
inventory buffer already exist, they do not run.