Blindly add file local variables for lexical scope
Lexical scope seems to be the standard nowadays for writing elisp.

No guarantees that anything works yet: I'll figure out the bugs later (which is
why I'm putting this in a branch of its own and hammering out things here.
Currently, bugtesting is very annoying due to the need to restart Emacs every
time due to some but somewhere that makes NetHack crash the second time it runs.