Update nethack-compat.el
I added a rule for it in the Makefile, since I noticed it wasn't byte-compiling,
and it really is needed.

I also changed the Common Lisp library from cl to cl-lib, since using the old cl
seems to be (very) deprecated. The call for this require was inside an
eval-when-compile, which I removed.