cl-libify all files
In the previous commit (b615a66), I switched from the cl library to cl-lib
library. This was fine, but it broke in a lot of small places. To help with this
transition, I ran cl-libify on every file.

On nethack-keys, a run of cl-libify changed a let variable name (map) to cl-map,
which broke a lot of things. I manually switched this back, and no changes were
made on nethack-keys.el.