Fix crash when running NetHack a second time
When running nethack a second time, nethack.el would crash with an error that
Emacs would stumble into an error that it couldn't
switch-to-buffer-other-window(#<killed buffer>). In Spacemacs, the error shown
was (wrong-type-argument stringp nil). This was because Emacs was attempting to
switch to a buffer that didn't exist, in this case, the nh-map-buffer. While I
had put in a check to create a map and inventory window if they didn't already
exist, this function would not get triggered. In turns out that the check that I
used, `(not nh-map-buffer)' and `(not nh-menu-buffer-table)' didn't get
triggered, since nh-map-buffer would evaluate to #<killed buffer>. `(not
\#<killed buffer>)' evaluates to nil, and the consequent would not get run.

I fixed this by using `(not (buffer-live-p nh-map-buffer))', but this also
didn't work. It turns out that nhapi-create-map-window and
nhapi-create-inventory-window already have a check so that no duplicate buffers
are created. Changing the check there to buffer-live-p and simply calling
nhapi-create-map-window and nhapi-create-inventory-window without any wrapping
if would suffice.