Do not close other windows
Previously, opening NetHack would take over the entire screen, leaving no room
for any other ``work'' (say, an EWW buffer for browsing spoilers).  I found this
rather annoying, especially when having code (or a debugger) open on one window
split with the NetHack one.  This commit removes calls to `delete-other-windows'
in `nhapi-raw-print' and `nhapi-restore-window-configuration'.
`nethack-kill-buffers' was updated to kill the windows for the status and
message buffer by itself, using a dolist rather than what was originally
repeated `kill-buffer' statements.  In my Emacs configuration (though not a
default one), the raw print buffer would not be the exit buffer, instead, it
would be whatever buffer the NetHack process was started from. To fix this, a
`pop-to-buffer' was added at the end of `nh-sentinel'.