Remove trailing newlines from winlisp.c
This bothers me.

In other news, I have yet to fix the general horrendous formatting and alignment
in the file (and possibly, tackle the process for not using the objectively
terrible GNU C brace style). My knowledge of C isn't that great, and, when I
tried to do the fix, I got a load of compiler warnings and errors. Perhaps I'm
serially incompetent, but I think I'll come back to this later.