Relocate ChangeLog entries for winlisp.c into patch
Previously, I didn't realize that there was a ChangeLog file inside win/lisp,
and I wasn't updating it, instead putting my changes into the outer ChangeLog.
This means that I have a lot of cryptic entries without a scope, meaning that it
is difficult to identify where the change occurred.  This change moves all of the
old ChangeLog entries into to inner ChangeLog inside of the patch.  I reworded
the scopes, too, so changes are easier to hunt down.

Inside of /win/lisp/ChangeLog, I wrapped lines and deleted trailing newlines.
This is maybe one of the things that I will be Nazi about.