Print message to echo area when ATR_NOHISTORY is set
At some point in the jump to 3.6, a new text attribute bitmask was introduced,
ATR_NOHISTORY, which displays messages to the player without logging them to
history.  AFAICT, this is used for double-digit numeric prefix arguments, in
which a ``Count: 10'' message is displayed without logging it into history.  To
create a system to honor this bitmask, a lisp_cmd message-nohistory was added
that would print these messages to the echo area rather than clogging up the
message buffer.  Note that in the ``count'' case, Emacs already puts the message
``C-u 10'' into the echo area, and after the action, a ``Count: 10'' gets put
into the echo area, which disappears after the move.

In addition to this change, I fixed the indentation of lisp_putstr.  However
much I hate GNU style of C, it is the way that the code is coded.