Update AUTHORS and README, now README.MD
I found that some of the information in these files was (very) out of date.  In
the case of INSTALL, it was still written in outline-mode (I don't think Org
mode was that well known yet?).

To fix these errors, I fleshed out the list of contributors in AUTHORS, move
README to README.md and pulled text from the website in, and copied (and
rewrote) the content from INSTALL into the new README.md.