Create Makefile for build directory
Since its surprisingly difficult to call a patch command from within Emacs and
smoothly transition the build to a general Makefile, we need to create a
Makefile inside the build directory.  This Makefile assumes the existence of a
directory `build/nethack`, which is where the NetHack source should be untarred
to.

Ideally, we will be able to start a build process, from within Emacs, which uses
this Makefile to patch and make NetHack.  Untarring will occur from within
Emacs, since we need to know the version of NetHack to know the name of the
tarball and source folder.

One alternative possibility to consider, if this solution doesn't work, is to
wget the NetHack source to a different filename, say `nethack.tar', and have all
of the other commands work off of that.  The only place where the NetHack
version would need to be known is the download (wget) command itself.