Remove nethack-tar and assume tarball at nethack.tgz
Writing the commit message for the last commit, I realized that it would be
easier to assume by default that the NetHack source would be contained within
nethack.tgz.  Keeping this in mind, I removed the function for nethack-tar, and
substituted it directly into the `nethack-download-nethack` function.
Similarly, I edited the definition for `nethack-untar-nethack` to blindly untar
nethack.tgz, rather than nethack-tar.

Since I added in a Makefile in the last commit which applies a patch, I modified
the command which calls this Makefile inside of `nethack-build`.  Now it calls
the commands `make patch`, `make hints` (or, sometimes `make hints-3.6`), and
`make build`.  These correspond to the build/Makefile.

I have yet to actually install the procedures `nethack-untar-nethack` and
`nethack-download-nethack` into `nethack-install`.  I'm not sure how to approach
the variables to pass to `nethack-install`, and I'll need to think over things
before I make any changes.