Create and use nethack-query-for-version
I had previously put this into to source code under the principle of wishful
thinking.  Now that its time to make the actual changes, I added in a function
which uses `read-answer` to query the user for the version.

I'm not the most familiar with Elisp right now, and so I'm not sure whether the
way I've bound the short-answer characters is proper or not.

Another problem that may crop up in the Elisp is my use of setq-default to
change the `nethack-version`.  Again, this may be completely violating standard
use, but I wouldn't know.