Create option no-download-p for nethack-install
When testing nethack-el, it is very useful to have an option to disable
downloading of the NetHack tarball.  The tarball is often large, and on my
network speeds, takes a minute or two to download -- an unacceptable amount of
time.  To fix this, I added a switch to skip the download of NetHack, instead
using the preexisting nethack.tgz.

As it stands, the code lacks abstraction.  Changing the nethack variable
beforehand doesn't guarantee the version to be installed, and changing the
install location is guaranteed either.