Change let to let* in nethack-download-nethack
Inside of `nethack-download-nethack`, a let should've been a let* to correctly
determine the URL to download NetHack from.

I couldn't find a way to separate out these two commits, but I also changed
`nethack-tar` inside of the aforementioned let*.  Adding a `/' in front of the
`nethack-tar` allows `url-copy-file` to download the NetHack tarball.