Untar actual NetHack tarball
Previously, the command to untar inside of `nethack-untar-nethack` did not
locate the correct location of nethack.tgz.  Since this command is run from the
nethack-el directory, we needed to prefix the location of nethack.tgz with
``build/''.

Note that this does not make for very good abstraction.  Perhaps
`nethack-untar-nethack` should take an argument of where to locate the NetHack
tarball.