Rename `nethack-build' to `nethack-build-program'
When copying from pdf-tools, I accidentally renamed the
`pdf-tools-build-program' to `nethack-build' rather than
`nethack-build-program'.  The bug that this leads to has two solutions.  While I
could have renamed the call of `nethack-build-program' to `nethack-build', I
instead renamed the defun instead.  This comes with the benefit that in the
future code I copy from pdf-tools, any reference to `pdf-tools-build-program' is
easily and readily renamed.