Fix compilation-cmd in nethack-build-program
These commits, while logically distinct, are hard to separate due to how close
they are together.

The current Makefile in the build directory takes as an ``input'' the
environment variable NH_VER_NODOTS to determine the patchfile.  This change
prefixes `make patch' with `NH_VER_NODOTS=', followed by the result of
`nethack-version-nodots`.

Furthermore, to correctly run the build Makefile, the build directory needs to
be set with the `-C' flag to make.  In this case, the build directory is
`build'.

As a last note, the check to apply the the hints/linux-lisp file through
setup.sh was done incorrectly before.  Instead of passing `nil' and `-1' to
`substring`, nethack-el passed `0' and `-1'.  This would trigger an error, which
this commit fixes.

Note that the build directory, `./build' is hardcoded here.  This leads to bad
abstraction, and should possibly be changed in the future.