Rename build/nethack to build/nethack-src
Since the executable NetHack will be installed at build/nethack, there cannot
also be a folder there.  This causes problems, especially at the `make install`
stage of the NetHack installation phase.

To accommodate this change, all references to the folder build/nethack need to
be renamed to build/nethack-src.  At this stage was when I most regretted not
implementing enough abstraction barriers earlier.  This commit diff can serve as
a good place to start replacing instances of the name of the build directory.