Remove skip-dependencies-p and force-dependencies-p
The pdf-tools that these functions were copied from utilizes automake/autoconf
in its make process, passing the flags `-d' or `-D' to the shell script to
download dependencies.  The build process for nethack-el does not use such a
process, so users will need to install their own dependencies.  This commit
removes all instances and references to either `skip-dependencies-p` or
`force-dependencies-p` in nethack.el.