Remove nethack-installer-os
Since pdf-tools, of which much of the installation code is copied from uses a
autoconf/automake in its build process, it had a need to determine what OS it
was being building from and pass said option to its make shell-script.  However,
nethack-el does not use such a build process, and I don't really have a plan to
implement such a process.  Therefore, `nethack-installer-os` is an unused
function.