Update compilation-cmd and pass a PREFIX to make
These several commits were intertwined in this one stanza of code, and it was
difficult to separate them out.

The old hints/linux-lisp did not correctly place the nethack program into the
build directory, instead opting to place it in ~/.local/bin, following the XDG
specification (kind of).  This is not a good thing, as it enforces my personal
preferences on whoever uses the hints.  To fix this, and to better integrate the
hints file into the build process, the hints file takes an ``input'' of the
environment variable PREFIX, which must be passed to it whenever we run `make
install`.

The new `let` statement for `compilation-cmd` reflects this, passing PREFIX as a
environment variable to the `build` make instruction.

The new `let` statement also utilizes a `format` rather than a `concat`, which
both is more similar to the way strings were set up in pdf-tools from which I
used as a guide to make the installation code, as well as being more readable in
general.

Since the `concat` changed into a `format`, the `while` statement to check if
the NetHack version was greater than 3.6 no longer works, as `nil` doesn't count
very well as a string.  To address this, the `while` was replaced by an `if`,
with a empty string as its alternative.

As a final note, the .gitignore was updated to ignore the build/nethack
executable and build/nethackdir, in which NetHack files are stored.

Note that in the current stage, NetHack does not run since it is ``Unable to
open SYSCF_FILE.''  This is likely a permissions error.