Use build-directory uniformly
This commit was triggered by a lack of abstraction throughout the system.  If
Emacs was not started from within the `nethack-el` directory, it would seek a
`.build` directory from wherever it was started.  This was unacceptable.

When cleaning out for abstraction, I noticed that the way the build process is
currently set up does not require `nethack-identify-build-directory` or
`nethack-locate-build-directory`, since the build directory contains its own
Makefile so that the Elisp half of things can easily build for different
versions.

Another change to note is that the call of `nethack-untar-nethack` was moved
from `nethack-install` to `nethack-build-program`, since now,
`nethack-untar-nethack` takes the build directory, not the target directory, as
an input.  A possible increase in abstraction and code-quality may come from
removing these hardcoded references to `build/nethack-src`.

Note, too, that `nethack-untar-nethack` creates `build/nethack-src` if it does
not yet already exist.