Let tar and patch fail silently
During testing, rebuilding the NetHack program over and over again means that
sometimes, `nethack-el` attempts to patch a folder that has already been
patched.  When this triggers an error, the entire make/build process fails.

To prevent this, I used a trick from https://stackoverflow.com/a/52967701 to
have patch fail quietly on errors and passed the `-r-` flag to it, disabling
rejects files.

A similar modification was done to the tar step of the process, passing the flag
`--ignore-command-error` to tar.  Since this change would break the 80-character
limit, I added another `%s` into the `format` that contains both
`--strip-components=1` and `--ignore-command-error`.