Add back in short answers to read-answer
My old commit which changed the values `?6` and `?3` to `nil` may not have done
the right thing.  It turns out that `read-answer`s behavior is dependent on the
variable `read-answer-short`, and I just happened to have had it set to `nil`.
This means that when I tested the install process on a cleaner Emacs install,
Emacs would error with the complaint that `nil` is not a sequence.