Fix download by using nethack-directory
Previously, `nethack-download-nethack` used "build/nethack.tgz" as the location
to download to.  This means that if Emacs was run from the user's home
directory, it would attempt to download to `~/build/nethack.tgz`.  To fix this,
we `expand-file-name` with the `nethack-directory` as the `default-directory`.

I didn't catch this bug earlier since this wasn't an aspect of initialization
process I was testing, as I always used `(nethack-install nil t)` or
`(nethack-install t)` before the reordering of arguments.