Remove references to executable in install process
Previously, the `nethack-install` function would error, complaining that
`executable` did not exist as a variable.  This is because the `pdf-tools.el`
that much of the installation code was copied from used cmake as a build tool,
and the exit message that cmake gave, along with the `nethack-program` variable
were used to determine if the program was successfully compiled.  However, I am
not using cmake, and the error message cannot be so used.  Furthermore, I do not
allow using `nethack-build`'s `target-directory` to influence the location of
the build process, perhaps a terrible choice, but it holds together so far.

To better handle the build process, we simply stat the existence of
`nethack-program` to determine if building succeeded or failed.  This does not
really work when updating NetHack, but when that occurs, usually, the entire
`nethack.el` directory should be nuked (you are using some quelpa recepie,
right?).