Blindly restore extended command functionality
I decided to start tackling the extended commands feature, which was removed
quite some time ago.

This commit blindly restores the functions that were removed.

One change I needed to make to the call of `nhapi-choose-attribute` in
`nhapi-get-ext-cmd` was passing a third empty string parameter.
`nhapi-choose-attribute` now takes a third argument of how to `quit`, and, not
knowing any better thing, I put the empty string.

It should also be noted that the commands previously prefixed with
`nethack-api-*` are now prefixed with `nhapi-*`.