Add keybinding `#` for extended command
Following the pattern in the other keybindings, a `nethack-command` needs to be
defined once in `nethack-cmd.el` and again in `nethack-keys.el`.

I followed the direction of `cmd.c` and gave the `cmdstr` for
`nethack-command-read-extended` as `#`.  Looking closely at things, `winlisp.c`
seems to define its own `cmd_index_t` ``taken from cmd.c'' that doesn't really
seem to do anything.  Perhaps it was a reference for writing `winlisp.c`?  I am
considering to remove it.

I'm starting to get a better idea of how the system is organized, both on the
Elisp and C sides of things.