Add # to cmd_index
This adds `#` to the `cmd_index` in `winlisp.c` so that extended commands are
even possible.  Contrary to my previous remark, `cmd_index` is very useful.  It
is used to translate incoming command strings to keystrokes.

However, since 3.4.3, `cmd.c` has changed quite a bit; `cmdlist` doesn't exist
anymore, and `extcmdlist` now also stores the key that corresponds to a command.
This means that it may be possible to leech off of `extcmdlist` to do the
string-character translation.

It should also be noted that I went and fixed the indentation and formatting of
`cmd_index`.  Less stress on my brain.