Remove build Makefile
Two functions from the build Makefile (`clean` and `nuke`) were reincorporated
into the master Makefile as `build-clean` and `build-nuke` respectively.

We can remove the `build/Makefile` now that things have been successfully
integrated as pure elisp functions.  I might write `clean` and `nuke` as elisp
functions in the future as elisp rather than Makefile, but right now, all of the
`nethack-build-` functions rely on the `build-directory` variable, which isn't
`custom`izable but is rather an argument passed to `nethack-build`.  I might
consider changing this in the future (and also move things into the elisp)..