Make target-directory in nethack-build optional
Also deletes the `target-directory` parameter passed to `nethack-build` from
`nethack-install`.

Also completely removes the `cl-check-type`.  I don't think this interface will
be exposed, and even if it is, is a `cl-check-type` really the best way of doing
things?

The relevant `setq`s and such has been moved into a `when target-directory`
sexp, so there shouldn't be too big of problems.

I also don't program elisp well, so I'm only half sure this is the right way to
use a `setq-default`.