Add file nethack-options with license
The barest of bones a file can be.

Thinking of making the first thing I implement is a check for the IBM-style
option.  It's useful and I think it would be nice to load a translation
table *when it needs to*.

I feel like it would be a bad idea to replace the options file altogether the
way it is currently written --- a bunch of the map-drawing stuff isn't
abstracted away but is actually writing a character to a place.  This would be
bad, say, if someone wanted to make a boulder a `0` or something like that.  Far
easier to just look for options that might screw things up (like IBM graphics)
and adjust for them as necessary.  At the very least, writing this lays the
groundwork in case the options file does need to be read in entirety.