Add function to actually compile NetHack
I don't know in what absent state-of-mind I was in when I forgot to copy this
over, but now, `nethack-build-compile` is an actual function now that is linked
into the rest of the `nethack-build` process.

It takes an argument, `callback`, which is passed to it through `nethack-build`.
Like `nethack-build`, it returns the `current-buffer`.