Make nethack-options-parse use symbols
As opposed to strings.  This might be a really bad idea, since string and symbol
matching goes pretty badly, but it enables the generation of special handles
akin to those already in `nethack-api`.  Maybe I should move the `intern` step
to when the code actually parses the `hilite_status`, though.