Create constant nethack-options-fields
I don't know how much this changes between versions, so it might not be a good
idea to make it a constant, but it makes sense to have it so checking is easier.
Also, I'm reversing the decision of the previous commit and storing the
options-fields as strings (for now...).