Move and expand nh-attr-face in nethack-options.el
Moves `nh-attr-face` to `nethack-options.el` and adds the necessary `require` in
`nethack-api.el` so things don't break.

`nh-attr-face` is also expanded upon so its argument can be either a string or a
symbol.  It automatically handles the presence (or abscence) of an `atr-`
prefix, as well as handling the aliases `normal` and `underline` for `none` and
`uline` respectively.