Add menucolor support
OK this is a big one (for me at least).  This adds support for what I think is
all of the required functionality for `MENUCOLOR=`.  A bunch of new things were
added to `nethack-options.el`, which is where most of the changes live.  Note
that the first time I ran it, `nethack` hanged for a while before I `C-g`d out,
so there might be something to do with the way I ran/tested it out of a REPL.

In `nethack-api.el`, `nhapi-add-menu` was modified to run
`nethack-options-highlight-menu` (contingent on `nethack-options-set-p`, both
new functions) to set the highlight for each function.  In a way, this kind of
makes it a hook.

To store the menucolors `nethack-options-menucolors` was defined which is
written to by the new function `nethack-options-get-menucolors`.  This trawls
through `nethack-options` to get the attributes set.

`nethack-options-color-face` was defined, which functions like
`nethack-attr-face`.  A little bit of new stuff was done, as the `pcase` uses
`rx` rather than just string matching (and there's still probably a better way
to do it) but it works for now.

When it is time to turn the property sexps into actual faces,
`nethack-options-attr-propertize` is what figures out whether to use
`nethack-options-color-face` or `nh-attr-face` depending on whether it is a
color argument or a text attribute.

Finally, the most important one (and the one which isn't written by me),
`nethack-options-highlight-menu`, mostly lifted out `nethack-example.el`.