Use CFLAGS SCORE_ON_BOTL and EXP_ON_BOTL
This fixes Xp and Score in the status window.  The status line didn't work
correctly before, as these would not update by default.  Instead, it is better
to (within the C) assume that we compile with all the status menu flags and then
handle whether they actually show up in the Lisp half.