Remove looped call to nhapi-print-status
In the C, when `BL_FLUSH` is passed as the `idx` to `lisp_status_update`, the
lisp command `nhapi-print-status` is run --- from the C half of the code!  It
would be redundant to call it again from the Lisp, especially since I'm
considering to move the age increment for status items to the
`nh-propertize-attribute` and so every time the status is printed, the age is
incremented.