Make nh-status-attributes alist
Also makes `nh-status-conditions` an alist.  Intead of relying on the obarray,
we use Lisp arrays!  I'm not sure if this is faster or slower, but I'm hoping it
makes it so that the attributes and conditions all feel a little less "hard
coded in" (though I'm realizing that doesn't actually mean a lot, given that
this is Lisp).

`nhapi-status-condition-update` has been accordingly modified, as well as a new
function, `nhapi-status-update`.  Both of these are called from the C.

The 3.6.6 patch has been modified to call these new functions, as well as to use
a new `lisp_status_update` routine instead of `genl_status_update`.  The old
`generate_status_line` is still hanging around, but commented out.  I'll remove
it once this whole operation finishes up nicely.

The `lisp_status_update` figures out whether it needs to call
`nhapi-status-condition-update`, `nhapi-status-update`, or `nhapi-print-status`.
All of the values are passed as strings (!) to the Lisp half.