Write substitute conditions function
This links up to the modification to `nethack-options-parse-hilite-status` in
which conditions are actually parsed correctly.

`nethack-options-substitute-conditions` reads in a behavior and makes the
substitutions turning it into a list of possible matches.  We should probably
take a look at the reference C code to see how they handle things like the `or`
I've set up here.