Add nethack-options-status-hilite
This new function returns a list of functions given a status string.

I've decided that when it comes to propertizing attributes, the code should call
through a list of possible functions.  This makes it so that we can parse
through the status hilites in `.nethackrc` a little easier (though the code is a
little hairy).

This commit also adds `nethack-options-status-function`, which once
`nethack-options-status-hilite` confirms that an element in
`nethack-options-hilites` should match, generates a function that takes inputs
`new`, `old`, `percent`, and `age`, and does the case checking.

This does mean that every time the hilites are drawn, two different checks are
made (one to get the function and one from the function itself).  We should
probably figure out how to memoize the results from
`nethack-options-status-hilites` to save on one of these checks.