Remove flatten-tree in options-status-hilite
Removes the call to `flatten-tree` in `nethack-options-status-hilite`, as it
flattens the tree too much, breaking open closures.

It turns out the tree-opening only needs to be done in one place: where each of
the two hilite cases has its own behavior.  Instead of returning the results as
a list, this commit makes it so that the results are returned through a nested
lambda which lists the result of each internal funcall.

This commit also flattens the structure of the code a little.