Only do second status-hilite condition if first fails
This (I think) matches the behavior of the C status-hilites code more closely.
I'll need to check, but I think I've got it right.  If not, this commit can be
reverted.

Instead of always doing two different checks and listing the results together
for a second clause, now, `nethack-options-status-hilite` checks the first,
and *if it fails* uses the second condition as the first's else clause.