Canonify nethack-status-buffer-format
By "canonify", I mean bringing the user-facing options closer to the underlying
C data structures.  Maybe this is a bad idea, but this certainly makes the code
a little more nicer (at least on the C side of things).

This commit updates `nh-status-string` to propertize things it touches.  The
`nh-status-char-to-format` is a big `pcase` taking in a possible
character (passed from `nethack-status-buffer-format`) and returns a `cons` cell
of the name of the attribute (for use with propertizing) and a format.

`nhapi-print-status` was updated to increment the age of attributes, since
`nhapi-update-status` is no longer a thing.  It handles incrementing both the
age in `nh-status-attributes` and `nh-status-conditions`.

As the old obarray-based searching for a function to render in the status bar
has been phased out, and some of my names are a little bit different from the
old ones, `nethack-status-buffer-format` and `nethack-status-header-line-format`
have been updated accordingly.

HD should only show when polymorphed, but I'm not storing that data on the Lisp
side of things right now.  Since I've got no check for it, the easiest way I can
think of to make it just work is to not display HD if it's value is 0.