Assume attribute values are strings
A bunch of bugs have been cropping up due to assumptions that the `new-value` is
a number, when it's actually a string.  Similar problems happen when it's a
string and `string-to-number` is called upon it.

As a future goal, I should start moving things one layer of abstraction down as
to avoid these types of errors.  In the meanwhile, however, I've set all of the
default `nh-status-attributes` in `nh-reset-status-variables` as strings of
zero, rather than numeric zeroes.

Similarly, the check for the new string format has been moved into the `let*`
statement within `nh-propertize-attribute`, so that if a string is of the new
format, it's handled correctly as a number.

This also fixes the bug that a >18 strength doesn't get the `Str:` prefix on it.