Prevent memory leak in nethack-options-status-hilite-results
This might never clear, which isn't a good thing.  Now, yes, there are way more
bulky lisp settings that continue to float around at program exit.  But this is
a hash table, and I think it would be best practice to clear it on exit (by
adding it to `nhapi-end`).

This also makes debugging easier --- now we are guaranteed to reread the nethack
status-hilite table to get new functions each time!