Remove memoizing nethack-options-status-function
`nethack-options-status-hilite` previously memoized the results of getting the
`nethack-options-status-function` in the hashtable
`nethack-options-status-hilite-results`.  As these lambdas were applied by
`nh-propertize-attribute`, their closure environments were bound to newer and
newer `attr`s.  As the lambda passed back had the same elemnts, binding/calling
the frame enviroments would stay until the next lambda call.  This is an
unacceptable issue.

I've removed the hashing, since otherwise it would crash with a circular loop
when it tried to `mapcan` the faces together.

Right now, it works to an acceptable extent.  I'm thinking about doing the
parsing as the .nethackrc is read, rather than when generating faces.