Make C half understand lisp version number
A new switch is added to `makedefs.c` so that asking for a `nethack --version`
tells you which lisp version the C code is meant to work with.

Important to note is that ccache doesn't quite handle this the right way, but I
don't also want to add `include/global.h` as a file to be diffed using
`mkpatch`.  Maybe I'll do that, though...