Make nethack-installed-p check both NetHack and Lisp versions
This commit first moves the compilation flag for `LISP_VERSION` from the
`linux-lisp` hints file into the proper `patchlevel.h`.  This means that
`ccache` will actually be able to detect when to recompile.

I'm not sure if this is the right move, as it will necessitate a recompile after
every version bump.  What we can do, is increment the minor version when the
C api changes, and keep incrementing patchlevel.  We could only check if the
major and minor versions of the lisp-patch is correct.

This commit also adds `patchlevel.h` to the `mkpatch` program.

And I've implemented the check for both NetHack and Lisp versions in
`nethack-installed-p`.

`nethack-el-version` has been bumped to `0.11.0`.