Fix extended commands
Adds back in use of `extended_cmd_id` which was removed earlier since I didn't
know what I was doing.

Before this commit, any extended command (that is, any command that `winlisp.c`
thought was a `CMD_EXT`) would trigger the completion, even if it were run
directly.  For example, pressing `M-a` wouldn't adjust, but instead would open a
completion window as if `#` were pressed.

Before this commit, `extended_cmd_id` didn't actually do anything, since it was
meant to be the return value of `lisp_get_ext_cmd`.  When I first added in
extended command support, I deleted this return line.

To fix this, `lisp_get_ext_cmd` now checks to see if `extended_cmd_id` is set.
If it is, then no prompt is given, it's value is reset to `0`, and the command
is returned.  If no `extended_cmd_id` has been set, only then it prompts for
completion.