Make IBM graphics work
Still working on DEC graphics.  IBM graphics are easier since there's
already (included in Emacs) support for decoding between cp437 and whatever
Emacs uses.  However, DEC graphics (what would be cp1090) is not included with
Emacs.

I'm not sure how efficient the new include with nethack-compat is.  In any case,
if it works, it works.