Make DEC graphics work
Mostly handled through a hand-coded translation table
(I hope that there's a better way to do this, but I'm not clever enough to
figure it out).

I'm not so sure about how efficient handling this in Lisp vs in C is, but it
works well enough to be usable.  Emacs doesn't have a character set to encode
cp1090/vt100 graphics, so I used a translation table that I took from
http://fileformats.archiveteam.org/wiki/DEC_Special_Graphics_Character_Set
(accessed 2021-04-23).

The logic that I used to determine whether I needed to render with high-bit set
lower is taken from `g_putch` in `wintty.c`.