Add nethack installed check for `nethack`
This changes `nethack-installed-p` so that it checks if `nethack-program` is
even an executable before it tries to run it as one, incase, idk, you're
starting from a blank install.

`nethack` is also changed so that `nethack-installed-p` is actually run to see
if nethack is installed before it tries to start a nethack process.  It will
also now automatically call a `nethack-install` if it isn't installed.

Note that `nethack-install` now has a new parameter `launch-nethack-p`, which is
how `nethack` is still run if nethack needs to be installed first.