Call nh-attr-face with full face name
The heart of the change is in the C-half, in the definition of `attr_to_string`,
each face name is given in full and passed as a symbol to `nh-attr-face`.  This
means that `nh-attr-face` (within `nethack-options.el`) can be changed to, by
default, process the symbol name as a face and return the `symbol-value`.  I've
left the function itself here for the sake of backwards compatibility with
previous versions of the lisp patch (not that this really matters while the
major version is 0).

In `nethack-api.el`, I've removed a call to `nh-attr-face` from
`nhapi-raw-print-bold`, instead opting to directly get a
`nethack-atr-bold-face`.

The Lisp version has also been bumped to 0.13.0.  The earliest compatible
version is still 0.12.0.