Make nh-atr-face correctly return a symbol
Because fonts are symbols, not variables as their treatment may sometimes
suggest, or at least, was suggested to me.  I didn't understand the way that
text properties would later be added, so I very much misunderstood the way emacs
handles fonts.  But no longer!

`nh-atr-face` now correctly returns a font as a symbol, rather than attempting
to evaluate it as a variable and miserably failing.  I've had to make a similar
change to `nhapi-raw-print-bold`, wherin `nethack-atr-bold-face` was evaluated
as a variable rather than being treated as a symbol.

I would like to make it known that I spent an unreasonable amount of time
looking in `winlisp.c`, and trying to blame through `enh-366.patch` in different
places to try to find if the problem was with quoting.  After all, at least from
the way the debugger presents it, it's hard to tell when something's a quote or
a variable already evaluated.  Whatever.  Now I don't need to bump the version.

I think that this error had slipped by me for so long because on my main laptop,
I had never actually tested the code beyond just assuming that it would work.
On my little thinkpad that I'm editing things on right now, I had to recompile
the whole thing to figure out where I had left off.  This meant that in starting
from ground 0, the little instances of backwards compatibility washed off.  For
you see, I had been playing `nethack.el` on this laptop fine for some time, with
a lisp version past 0.13.0.  However, what I didn't realize was that my
`init.el` had been setting my `nethack-program` to `../nethack`, rather than the
`./build` directory.  The executable, then, was passing symbols that were *not*
fonts to the Lisp half, meaning that the non-font symbols would pass right
through the then-forms and straight into the else.  But I have things fixed now.

As an aside, it seems like the later it gets, the needless verbosity of my
commit messages increases.  This commit is like literally two lines of change.